Ajout de la creation et validation d'une request update

+++ GameTrip/GameTrip.Domain/Entities/RequestLocationUpdate.cs
@@ -7,7 +7,7 @@ public class RequestLocationUpdate
     public string? Name { get; set; }
     public string? Description { get; set; }
     public decimal? Latitude { get; set; }
-    public double? Longitude { get; set; }
+    public decimal? Longitude { get; set; }
     public ICollection<Picture>? Pictures { get; set; }
     public ICollection<Game>? Games { get; set; }
 }
